Stop indefinitely caching loaded packages
This is a performance hit in some cases but seems worth it.

It's a bit awkward to have the WeakRef to break the circular
dependency, especially since no global variables are defined.

To match the pkg-config implementation, the above is staying in.
This can be removed after the dust has settled more and there's
still no need.